requirements.txt
# Core Framework
Flask==2.3.3
python-dotenv==1.0.0

# Enhanced Audio Processing with MFCC
librosa==0.10.1
sounddevice==0.4.6
pyaudio==0.2.11
numpy==1.26.4
scipy==1.11.4

# Enhanced Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.14
Pillow==10.1.0

# Machine Learning - Enhanced
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.35.2

# Facial Emotion Detection
deepface==0.0.79
tensorflow==2.13.0

# Enhanced Audio Features
librosa==0.10.1
soundfile==0.12.1

# LLM Integration
requests==2.31.0
ollama==0.1.7

# Utilities
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.7.5
threadpoolctl==3.2.0